claude_code_sdk/pyproject.toml
[project]
name = "cc-sdk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk>=0.0.20",
    "ipykernel>=6.29.5",
    "mcp-server-git>=2025.1.14",
    "python-dotenv>=1.1.1",
]
